[Inductor Cutlass backend] Tolerate dynamic shapes
Previously, when the Cutlass backend was
enabled, using dynamic shapes could lead
to exceptions during JIT.

With this change, there are guards in place to just disable the Cutlass backend if dynamic dimensions are involved.

In addition, if no choices for a GEMM are available using the selected backends,
then an ATen Kernel is used as fallback,
even if the ATen backend is not enabled.

Test:
CI
Additional unit test in test_cutlass_backend.py

ghstack-source-id: 1cff352b70b3757942e8f2007bfad991ac014759
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121497